Reverted Giphy renaming for packages
