+++ WordPress/src/main/java/org/wordpress/android/modules/ApplicationModule.java
@@ -10,8 +10,8 @@ import org.wordpress.android.ui.accounts.signup.SettingsUsernameChangerFragment;
 import org.wordpress.android.ui.accounts.signup.UsernameChangerFullScreenDialogFragment;
 import org.wordpress.android.ui.domains.DomainRegistrationDetailsFragment.CountryPickerDialogFragment;
 import org.wordpress.android.ui.domains.DomainRegistrationDetailsFragment.StatePickerDialogFragment;
-import org.wordpress.android.viewmodel.gifs.provider.GifProvider;
-import org.wordpress.android.viewmodel.gifs.provider.TenorProvider;
+import org.wordpress.android.viewmodel.giphy.provider.GifProvider;
+import org.wordpress.android.viewmodel.giphy.provider.TenorProvider;
 import org.wordpress.android.ui.news.LocalNewsService;
 import org.wordpress.android.ui.news.NewsService;
 import org.wordpress.android.ui.reader.ReaderPostWebViewCachingFragment;
