+++ WordPress/src/main/java/org/wordpress/android/ui/giphy/GiphyMediaViewHolder.kt
@@ -14,7 +14,7 @@ import org.wordpress.android.util.getDistinct
 import org.wordpress.android.util.image.ImageManager
 import org.wordpress.android.util.image.ImageType.PHOTO
 import org.wordpress.android.util.redirectContextClickToLongPressListener
-import org.wordpress.android.viewmodel.gifs.GiphyMediaViewModel
+import org.wordpress.android.viewmodel.giphy.GiphyMediaViewModel
 
 /**
  * Represents a single item in the [GiphyPickerActivity]'s grid (RecyclerView).
@@ -75,7 +75,7 @@ class GiphyMediaViewHolder(
      * Update the views to use the given [GiphyMediaViewModel]
      *
      * The [mediaViewModel] is optional because we enable placeholders in the paged list created by
-     * [org.wordpress.android.viewmodel.gifs.GiphyPickerViewModel]. This causes null values to be bound to
+     * [org.wordpress.android.viewmodel.giphy.GiphyPickerViewModel]. This causes null values to be bound to
      * [GiphyMediaViewHolder] instances.
      */
     override fun bind(item: GiphyMediaViewModel?) {
