+++ WordPress/src/main/java/org/wordpress/android/viewmodel/giphy/CoroutineScopedViewModel.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.viewmodel.gifs
+package org.wordpress.android.viewmodel.giphy
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.CoroutineScope
