+++ WordPress/src/main/java/org/wordpress/android/viewmodel/giphy/GiphyPickerDataSource.kt
@@ -1,9 +1,9 @@
-package org.wordpress.android.viewmodel.gifs
+package org.wordpress.android.viewmodel.giphy
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.paging.PositionalDataSource
-import org.wordpress.android.viewmodel.gifs.provider.GifProvider
+import org.wordpress.android.viewmodel.giphy.provider.GifProvider
 
 /**
  * The PagedListDataSource that is created and managed by [GiphyPickerDataSourceFactory]
